Move LVM helpers to dedicated module
